adding check permission system, fixing the logger, and adding delete channel system (not done yet)
i hate myself, i think i'll kill myself one of these days

+++ src/nuke/nuke.js
@@ -1,4 +1,8 @@
-import deleteAllRoles from "./role.js";
-export default function nuke() {
-  deleteAllRoles();
+import deleteAllChannels from "./channel.js";
+import { deleteAllRoles } from "./role.js";
+import checkPermissions from "./checkPermission.js";
+export default async function nuke() {
+  if (!(await checkPermissions())) return;
+  await deleteAllRoles();
+  await deleteAllChannels();
 }
